+++ quantforge/__init__.py
@@ -42,6 +42,7 @@ from .lookback import (
 from .heston import heston_price, heston_smile
 from .heston_mc import heston_qe_mc
 from .bates import bates_price, bates_smile
+from .double_heston import double_heston_price, double_heston_smile
 from .kou import kou_price, kou_smile
 from .cgmy import cgmy_price, cgmy_smile
 from .nig import nig_price, nig_smile
@@ -178,7 +179,7 @@ try:  # pragma: no cover - trivial availability branch
 except ImportError:  # pragma: no cover
     HAS_NUMPY = False
 
-__version__ = "1.121.0"
+__version__ = "1.122.0"
 
 __all__ = [
     "OptionType",
@@ -227,6 +228,8 @@ __all__ = [
     "heston_qe_mc",
     "bates_price",
     "bates_smile",
+    "double_heston_price",
+    "double_heston_smile",
     "kou_price",
     "kou_smile",
     "cgmy_price",